third_party/WebKit/LayoutTests/crypto/subtle/aes-cbc/decrypt-failures-expected.txt

Decrypting truncated AES-CBC ciphertext should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS: Decryption succeeded
error is: OperationError
PASS: decrypting failed. ciphertext length: 0
error is: OperationError
PASS: decrypting failed. ciphertext length: 79
error is: OperationError
PASS: decrypting failed. ciphertext length: 64
error is: OperationError
PASS: decrypting failed. ciphertext length: 1
error is: OperationError
PASS: decrypting failed. ciphertext length: 15
error is: OperationError
PASS: decrypting failed. ciphertext length: 16
error is: OperationError
PASS: decrypting failed. ciphertext length: 17
PASS successfullyParsed is true

TEST COMPLETE
